feat: Added list of search results
Search results are now shown in a list every time a search is done. This uses
the knockout observables. However, the binding is applied manually because
applying the bindings in initMap will interfere with the bindings by the
viewModel. So ko.applyBindings(new viewModel()); line in the app.js is removed.