feat: Added name and address info for yelp results
Yelp results now contain names and addresses. However, the map markers from
google results and yelp are separately dropped. This makes them not aware when a
marker from other API is clicked thus, the InfoWindow does not disappear. Google
results may actually be completely removed as yelp will be the primary
information provide for the markers.