feat: Made yelp search dynamic(partial)
Yelp now takes coordinates as parameter and use it as bias on search results.
Also added a radius parameter. The radius will be from the coordinate passed in
the function. While these are added, yelp can only take search terms such as
"pizza" or "burger" and not a specific address or location that the searchBox
autocompletes to.