feat: Foursquare search implemented
Users can now search single terms such as "pizza" or "burgers" and get results
with markers on the map. The results are also visible under the search bar and
responsive to clicks. If the user clicks on a place under the search bar, single
map marker will drop indicating the location of the place.