feat: Added search button and dropdown
The dropdown does not have any content yet. I will try to get the selections
from Foursquare dynamically. Search button is fully functional and usable. It is
provided so that users not only have the option to press enter to run a search
but also to tap the search button to search. Additional validation for
Foursquare data is also added.