fix: Removed undefined parts of addresses
Some places dont have certain parts of their address on the record (Foursquare
database) so they show up undefined. This is because I hard coded all the parts
of the address. In this commit, the addresses are now dynamically formatted.
This is with the use of "formattedAddress" member of the object that Foursquare
returns.