feat: Show InfoWindow by clicking on the list
All markers used to be removed when clicking on the list of places. In this
commit, markers are not removed and recreated but rather only shows an
infoWindow when its corresponding name on the list is clicked. This is done
because removing all markers and recreating one marker that corresponds to the
clicked place on the list is expensive and takes resources.