refactor: Eliminated extra 'places' variable
An observableArray variable 'places' is deleted. The information that it used to
contain is transfered to maps.markers variable, which is an observableArray
since last commit. This commit removes data duplication among these two
variables. Finally, an attemp in notifying observables through valueHasMutated()
is done. Right now, it does not work as expected.