fix: Results list does not update
The markers on the map update as a category is selected. But the clickable list
of all markers are still there and active. So when they are clicked, the
infowindows pop up even tho the marker is not visible on the map. I created a
new observable variable to help update the list independent to the markers on
the map because the list cant rely on the array that the map markers uses. This
is because it is no possible for markers' visibility status to change with the
use of observables.